+++ modules/ethereum/Cargo.toml
@@ -10,8 +10,8 @@ version = '0.9.0'
 [dependencies]
 rustc-hex = { version = "2.1.0", default-features = false }
 serde = { version = "1.0.101", optional = true }
-codec = {default-features = false, package = 'parity-scale-codec', version = '1.3.4' }
 sp-core = { version = "2.0.0", default-features = false }
+codec = {default-features = false, package = 'parity-scale-codec', version = '1.3.4' }
 frame-support = { version = "2.0.0", default-features = false }
 frame-system = { version = "2.0.0", default-features = false }
 pallet-balances = { version = "2.0.0", default-features = false }
@@ -20,13 +20,15 @@ clover-evm = { version = "0.9.0", default-features = false, path = "../evm" }
 sp-runtime = { version = "2.0.0", default-features = false }
 sp-std = { version = "2.0.0", default-features = false }
 sp-io = { version = "2.0.0", default-features = false }
-primitives = { version = '0.9.0', package = 'clover-primitives', path = '../../primitives'}
 evm = { version = "0.18", features = ["with-codec"], default-features = false }
 ethereum = { version = "0.5", default-features = false, features = ["with-codec"] }
 ethereum-types = { version = "0.9", default-features = false }
 rlp = { version = "0.4", default-features = false }
 sha3 = { version = "0.8", default-features = false }
 libsecp256k1 = { version = "0.3", default-features = false }
+fp-consensus = { path = "../../primitives/consensus", default-features = false }
+fp-rpc = { path = "../../primitives/rpc", default-features = false }
+fp-evm = { default-features = false, path = "../../primitives/evm" }
 
 [features]
 default = ["std"]
@@ -43,11 +45,13 @@ std = [
 	"clover-evm/std",
 	"sp-io/std",
 	"sp-std/std",
-	"primitives/std",
 	"ethereum/std",
 	"ethereum-types/std",
 	"rlp/std",
 	"sha3/std",
 	"libsecp256k1/std",
 	"evm/std",
+	"fp-consensus/std",
+	"fp-rpc/std",
+	"fp-evm/std",
 ]
