+++ modules/evm/src/runner/mod.rs
@@ -3,7 +3,7 @@ pub mod builtin;
 
 use sp_std::vec::Vec;
 use sp_core::{H160, U256, H256};
-use primitives::{CallInfo, CreateInfo};
+use fp_evm::{CallInfo, CreateInfo};
 use crate::Trait;
 
 pub trait Runner<T: Trait> {
