+++ node/Cargo.toml
@@ -25,6 +25,8 @@ jsonrpc-pubsub = "15.0.0"
 structopt = '0.3.8'
 serde_json = "1.0"
 
+pallet-contracts = { version = "2.0.0", default-features = false }
+pallet-contracts-rpc = { version = "0.8.0" }
 pallet-transaction-payment-rpc = { version = '2.0.0' }
 sc-basic-authorship = { version = '0.8.0-rc6' }
 sc-cli = { features = ['wasmtime'], version = '0.8.0-rc6' }
