+++ primitives/Cargo.toml
@@ -18,12 +18,7 @@ strum_macros = "0.19.2"
 int-enum = "0.4.0"
 enum-iterator = "0.6.0"
 sp-core = { version = '2.0.0', default-features = false }
-sp-api = { version = "2.0.0", default-features = false }
 sp-runtime = { version = '2.0.0', default-features = false }
-sp-std = { version = "2.0.0", default-features = false }
-evm = { version = "0.18", default-features = false, features = ["with-codec"] }
-ethereum = { version = "0.5", default-features = false, features = ["with-codec"] }
-ethereum-types = { version = "0.9", default-features = false }
 
 [features]
 default = ["std"]
@@ -33,10 +28,4 @@ std = [
   "codec/std",
   "sp-runtime/std",
   "sp-core/std",
-  "sp-std/std",
-  "evm/std",
-  "evm/with-serde",
-  "sp-api/std",
-  "ethereum/std",
-  "ethereum-types/std",
 ]
