+++ primitives/consensus/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "fp-consensus"
+authors = ['Clover Network <https://github.com/clover-network>']
+description = "Clover primitives definitions"
+edition = "2018"
+license = 'GPL-3.0'
+homepage = "https://substrate.dev"
+repository = 'https://github.com/clover-network/clover'
+version = '0.9.0'
+
+[dependencies]
+sp-std = { version = "2.0.0", default-features = false }
+sp-runtime = { version = "2.0.0", default-features = false }
+sp-core = { version = "2.0.0", default-features = false }
+codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false, features = ["derive"] }
+
+[features]
+default = ["std"]
+std = [
+	"sp-std/std",
+	"sp-runtime/std",
+	"sp-core/std",
+	"codec/std",
+]
