+++ primitives/evm/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "fp-evm"
+authors = ['Clover Network <https://github.com/clover-network>']
+description = "Clover primitives definitions"
+edition = "2018"
+license = 'GPL-3.0'
+homepage = "https://substrate.dev"
+repository = 'https://github.com/clover-network/clover'
+version = '0.9.0'
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+sp-core = { version = "2.0.0", default-features = false }
+sp-std = { version = "2.0.0", default-features = false }
+serde = { version = "1.0.101", optional = true, features = ["derive"] }
+codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false }
+evm = { version = "0.18", default-features = false, features = ["with-codec"] }
+
+[features]
+default = ["std"]
+std = [
+	"sp-core/std",
+	"sp-std/std",
+	"serde",
+	"codec/std",
+	"evm/std",
+	"evm/with-serde",
+]
